requirements.txt
python-telegram-bot==22.7
pandas
requests
schedule
yfinance
aiohttp
